Conflicts:
	Graph.cpp
	Graph.h

fixed Graph.h

+++ Graph.h
@@ -42,8 +42,8 @@ public:
   
 	//Return the cost of the edge from node1 to node2. If there
 	// is no edge, return -1.
-	int getCost(int node1, int node2);
-
+	double getCost(int node1, int node2);
+	
 	//Add an edge from node1 to node2, and from node2 to node1, with
 	// the given cost. If the cost is < 0, throw a string exception.
 	// If the edge already exists, just update the cost
